requirements.txt
Jinja2==3.1.2
MarkupSafe==2.1.1
PyJWT==2.5.0
SQLAlchemy==1.4.41
anyio==3.6.1
bcrypt==4.0.1
bezier==2021.2.12
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
fastapi-login==1.8.2
fastapi==0.85.0
greenlet==1.1.3
h11==0.14.0
idna==3.4
importlib-metadata==5.0.0
numpy==1.21.6
passlib==1.7.4
playwright==1.27.0
pyasn1==0.4.8
pydantic==1.10.2
pyee==8.1.0
python-ghost-cursor==0.1.1
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.4.0
uvicorn==0.18.3
zipp==3.9.0